Organize related files into kaji.command package
