--- src/main/java/ToDo.java
@@ -1,8 +1,0 @@
-/**
- * Represents a Todo task.
- */
-public class ToDo extends Task {
-    public ToDo(String type, Boolean isDone, String description) {
-        super(type, isDone, description);
-    }
-}
