+++ src/main/java/kaji/Kaji.java
