+++ src/main/java/kaji/KajiException.java
