+++ src/main/java/kaji/Storage.java
