+++ src/main/java/kaji/command/DeleteCommand.java
@@ -1,9 +1,22 @@
+package kaji.command;
+
+import kaji.KajiException;
+import kaji.Storage;
+import kaji.TaskList;
+import kaji.Ui;
+
 /**
- * Deals with the list command.
+ * Deals with the delete command.
  */
-public class ListCommand extends Command {
+public class DeleteCommand extends Command {
+    private int taskId;
+
+    public DeleteCommand(int taskId) {
+        this.taskId = taskId;
+    }
+
     /**
-     * Executes the list command.
+     * Executes the delete command.
      *
      * @param tasklist The task list to operate on.
      * @param ui The UI to interact with user.
@@ -12,7 +25,7 @@ public class ListCommand extends Command {
      */
     @Override
     public void execute(TaskList tasklist, Ui ui, Storage storage) throws KajiException {
-        ui.showTaskList(tasklist);
+        tasklist.deleteTask(taskId, ui);
     }
 
     /**
