+++ src/main/java/kaji/command/MarkCommand.java
@@ -1,15 +1,22 @@
+package kaji.command;
+
+import kaji.KajiException;
+import kaji.Storage;
+import kaji.TaskList;
+import kaji.Ui;
+
 /**
- * Deals with the delete command.
+ * Deals with the mark command.
  */
-public class DeleteCommand extends Command {
+public class MarkCommand extends Command {
     private int taskId;
 
-    public DeleteCommand(int taskId) {
+    public MarkCommand(int taskId) {
         this.taskId = taskId;
     }
 
     /**
-     * Executes the delete command.
+     * Executes the mark command.
      *
      * @param tasklist The task list to operate on.
      * @param ui The UI to interact with user.
@@ -18,7 +25,7 @@ public class DeleteCommand extends Command {
      */
     @Override
     public void execute(TaskList tasklist, Ui ui, Storage storage) throws KajiException {
-        tasklist.deleteTask(taskId, ui);
+        tasklist.markTask(taskId, ui);
     }
 
     /**
